+++ src/PaymentMeans.php
@@ -19,6 +19,7 @@ class PaymentMeans implements XmlSerializable
     private $instructionNote;
     private $paymentId;
     private $payeeFinancialAccount;
+    private $paymentMandate;
 
     /**
      * @return string
@@ -131,6 +132,23 @@ class PaymentMeans implements XmlSerializable
         return $this;
     }
 
+    /**
+     * @return PaymentMandate
+     */
+    public function getPaymentMandate(): ?PaymentMandate
+    {
+        return $this->paymentMandate;
+    }
+
+    /**
+     * @param string $paymentMandate
+     * @return PaymentMeans
+     */
+    public function setPayementMandate(?PaymentMandate $paymentMandate): PaymentMeans
+    {
+        $this->paymentMandate = $paymentMandate;
+        return $this;
+    }
     public function xmlSerialize(Writer $writer): void
     {
         $writer->write([
@@ -166,6 +184,12 @@ class PaymentMeans implements XmlSerializable
         if ($this->getPayeeFinancialAccount() !== null) {
             $writer->write([
                 Schema::CAC . 'PayeeFinancialAccount' => $this->getPayeeFinancialAccount()
+            ]);
+        }
+
+        if($this->getPaymentMandate() !== null){
+            $writer->write([
+                Schema::CAC . 'PaymentMandate' => $this->getPaymentMandate()
             ]);
         }
     }